Clean/db/requirements.txt
onnxruntime 
insightface 
faiss-cpu 
numpy 
opencv-python
pillow 
fastapi
uvicorn
pytest-shutil
python-multipart
